Move vasp.vasp*put to vasp.*puts. Also, maintain backwards compatibility with vaspio.vasp_*put
