+++ docs/pymatgen.io.feff.rst
@@ -1,0 +1,28 @@
+pymatgen.io.feff package
+========================
+
+Subpackages
+-----------
+
+.. toctree::
+
+
+Submodules
+----------
+
+pymatgen.io.feff.sets module
+----------------------------
+
+.. automodule:: pymatgen.io.feff.sets
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pymatgen.io.feff
+    :members:
+    :undoc-members:
+    :show-inheritance:
