+++ docs/pymatgen.io.vasp.rst
@@ -1,0 +1,52 @@
+pymatgen.io.vasp package
+========================
+
+Subpackages
+-----------
+
+.. toctree::
+
+
+Submodules
+----------
+
+pymatgen.io.vasp.GWvaspinputsets module
+---------------------------------------
+
+.. automodule:: pymatgen.io.vasp.GWvaspinputsets
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pymatgen.io.vasp.sets module
+----------------------------
+
+.. automodule:: pymatgen.io.vasp.sets
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pymatgen.io.vasp.vasp_input module
+----------------------------------
+
+.. automodule:: pymatgen.io.vasp.vasp_input
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pymatgen.io.vasp.vasp_output module
+-----------------------------------
+
+.. automodule:: pymatgen.io.vasp.vasp_output
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pymatgen.io.vasp
+    :members:
+    :undoc-members:
+    :show-inheritance:
