+++ pymatgen/alchemy/materials.py
@@ -26,7 +26,7 @@ from monty.json import MontyDecoder
 
 from pymatgen.core.structure import Structure
 from pymatgen.io.cif import CifParser
-from pymatgen.io.vasp.vasp_input import Poscar
+from pymatgen.io.vasp.inputs import Poscar
 from pymatgen.serializers.json_coders import PMGSONable
 from pymatgen.matproj.snl import StructureNL
