+++ pymatgen/analysis/diffusion_analyzer.py
@@ -32,7 +32,7 @@ import numpy as np
 from pymatgen.core import Structure, get_el_sp
 import pymatgen.core.physical_constants as phyc
 from pymatgen.serializers.json_coders import PMGSONable
-from pymatgen.io.vasp.vasp_output import Vasprun
+from pymatgen.io.vasp.outputs import Vasprun
 from pymatgen.util.coord_utils import pbc_diff
