+++ pymatgen/command_line/enumlib_caller.py
@@ -44,7 +44,7 @@ from monty.fractions import fractions
 
 from six.moves import reduce
 
-from pymatgen.io.vasp.vasp_input import Poscar
+from pymatgen.io.vasp.inputs import Poscar
 from pymatgen.core.sites import PeriodicSite
 from pymatgen.core.structure import Structure
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
