+++ pymatgen/io/tests/test_cif.py
@@ -9,7 +9,7 @@ import warnings
 import numpy as np
 
 from pymatgen.io.cif import CifParser, CifWriter, CifBlock
-from pymatgen.io.vasp.vasp_input import Poscar
+from pymatgen.io.vasp.inputs import Poscar
 from pymatgen import Element, Specie, Lattice, Structure, Composition, DummySpecie
 from pymatgen.analysis.structure_matcher import StructureMatcher
