+++ pymatgen/io/tests/test_cssr.py
@@ -18,7 +18,7 @@ import unittest
 import os
 
 from pymatgen.io.cssr import Cssr
-from pymatgen.io.vasp.vasp_input import Poscar
+from pymatgen.io.vasp.inputs import Poscar
 from pymatgen.core.structure import Structure
 
 test_dir = os.path.join(os.path.dirname(__file__), "..", "..", "..",
