+++ pymatgen/io/tests/test_xyz.py
@@ -19,7 +19,7 @@ import os
 
 from pymatgen.core.structure import Molecule
 from pymatgen.io.xyz import XYZ
-from pymatgen.io.vasp.vasp_input import Poscar
+from pymatgen.io.vasp.inputs import Poscar
 
 test_dir = os.path.join(os.path.dirname(__file__), "..", "..", "..",
                         'test_files')
