+++ pymatgen/io/tests/test_zeoio.py
@@ -19,7 +19,7 @@ from pymatgen.io.cif import CifParser
 from pymatgen.io.zeo import ZeoCssr, ZeoVoronoiXYZ, get_voronoi_nodes, \
     get_high_accuracy_voronoi_nodes, get_void_volume_surfarea, \
     get_free_sphere_params
-from pymatgen.io.vasp.vasp_input import Poscar
+from pymatgen.io.vasp.inputs import Poscar
 from pymatgen.analysis.bond_valence import BVAnalyzer
 
 try:
